Handled some issues
