+++ frontend/src/register_api.js
@@ -1,0 +1,11 @@
+export const register = (data) => {
+  const options = {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    mode: "cors",
+    body: JSON.stringify(data),
+  };
+  return fetch("http://127.0.0.1:8000/auth/users/", options);
+};
